Moved configureBookGrid to BookGrid
