+++ src/main/java/com/karankumar/bookproject/ui/shelf/BooksInShelfView.java
@@ -27,7 +27,6 @@ import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.textfield.TextField;
-import com.vaadin.flow.data.renderer.LocalDateRenderer;
 import com.vaadin.flow.data.value.ValueChangeMode;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
@@ -35,9 +34,6 @@ import com.vaadin.flow.router.RouteAlias;
 import lombok.extern.java.Log;
 
 import javax.transaction.NotSupportedException;
-import java.time.format.DateTimeFormatter;
-import java.time.format.FormatStyle;
-import java.util.Comparator;
 import java.util.List;
 import java.util.function.Predicate;
 import java.util.logging.Level;
@@ -160,83 +156,13 @@ public class BooksInShelfView extends VerticalLayout {
         throw new NotSupportedException("Shelf " + shelfName + " has not been added as a case in switch statement.");
     }
 
-
-    private String combineTitleAndSeries(Book book) {
-        if (book.existsSeriesPosition()) {
-            return String.format("%s (#%d)", book.getTitle(), book.getSeriesPosition());
-        }
-
-        return book.getTitle();
-    }
-
     private void configureBookGrid() {
-        bookGrid.asSingleSelect()
-                .addValueChangeListener(event -> {
-                    if (event.getValue() != null) {
-                        editBook(event.getValue());
-                    }
-                });
-
-        resetGridColumns();
-
-        addTitleColumn();
-        addAuthorColumn();
-        addGenreColumn();
-        addDateStartedColumn();
-        addDateFinishedColumn();
-        addRatingColumn();
-        addPagesColumn();
-        addPagesReadColumn();
-    }
-
-    private void addPagesColumn() {
-        bookGrid.addColumn(PAGES_KEY);
-    }
-
-    private void addPagesReadColumn() {
-        bookGrid.addColumn(PAGES_READ_KEY);
-    }
-
-    private void addRatingColumn() {
-        bookGrid.addColumn(RATING_KEY);
-    }
-
-    private void addDateFinishedColumn() {
-        bookGrid.addColumn(new LocalDateRenderer<>(
-                Book::getDateFinishedReading, DateTimeFormatter.ofLocalizedDate(FormatStyle.MEDIUM)))
-                .setHeader("Date finished reading")
-                .setComparator(Comparator.comparing(Book::getDateStartedReading))
-                .setSortable(true)
-                .setKey(DATE_FINISHED_KEY);
-    }
-
-    private void addDateStartedColumn() {
-        bookGrid.addColumn(new LocalDateRenderer<>(
-                Book::getDateStartedReading, DateTimeFormatter.ofLocalizedDate(FormatStyle.MEDIUM)))
-                .setHeader("Date started reading")
-                .setComparator(Comparator.comparing(Book::getDateStartedReading))
-                .setKey(DATE_STARTED_KEY);
-    }
-
-    private void addGenreColumn() {
-        bookGrid.addColumn(GENRE_KEY);
-    }
-
-    private void addAuthorColumn() {
-        bookGrid.addColumn(AUTHOR_KEY)
-                .setComparator(Comparator.comparing(author -> author.getAuthor().toString()))
-                .setSortable(true);
-    }
-
-    private void addTitleColumn() {
-        bookGrid.addColumn(this::combineTitleAndSeries) // we want to display the series only if it is bigger than 0
-                .setHeader("Title")
-                .setKey(TITLE_KEY)
-                .setSortable(true);
-    }
-
-    private void resetGridColumns() {
-        bookGrid.setColumns();
+        BookGrid bookGrid = new BookGrid(this.bookGrid);
+        bookGrid.configure(event -> {
+            if (event.getValue() != null) {
+                editBook(event.getValue());
+            }
+        });
     }
 
     private void updateGrid() {
